# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ipmoney API
  version: 0.1.0
  license:
    name: Proprietary
    url: https://ipmoney.com
  description: |
    Ipmoney 专利交易平台（小程序 + 用户 H5 + 后台）对外 API（P0）。

    业务约束（P0）：
    - 标的范围：仅 CN 专利（数据以用户上传/后台录入为准）
    - 资金路径：订金与尾款均线上支付并托管
    - 合同签署：线下签署完成后由客服/运营确认里程碑解锁尾款支付
    - 佣金：卖家承担；放款时从卖家应收中扣除
servers:
  - url: https://api.ipmoney.com/v1
...
security:
  - BearerAuth: []
paths:
...
  /public/artworks/{artworkId}:
    get:
      tags: [Artworks]
      summary: 获取书画作品详情（公开）
      operationId: getPublicArtworkById
      security: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtworkPublic"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /search/artworks:
    get:
      tags: [Search]
      summary: 书画专区检索
      operationId: searchArtworks
      security: []
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/ArtworkCategory"
        - $ref: "#/components/parameters/CalligraphyScript"
        - $ref: "#/components/parameters/PaintingGenre"
        - $ref: "#/components/parameters/Creator"
        - $ref: "#/components/parameters/CreationYearStart"
        - $ref: "#/components/parameters/CreationYearEnd"
        - $ref: "#/components/parameters/PriceType"
        - $ref: "#/components/parameters/PriceMin"
        - $ref: "#/components/parameters/PriceMax"
        - $ref: "#/components/parameters/DepositMin"
        - $ref: "#/components/parameters/DepositMax"
        - $ref: "#/components/parameters/RegionCode"
        - $ref: "#/components/parameters/ArtworkSortBy"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedArtworkSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /artworks:
    post:
      tags: [Artworks]
      summary: 创建书画草稿（发布方）
      operationId: createArtwork
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Artworks]
      summary: 我的书画列表（发布方）
      operationId: listMyArtworks
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkStatus"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedArtwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /artworks/{artworkId}:
    get:
      tags: [Artworks]
      summary: 获取书画详情（发布方/已登录）
      operationId: getArtworkById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Artworks]
      summary: 更新书画草稿（发布方）
      operationId: updateArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /artworks/{artworkId}/submit:
    post:
      tags: [Artworks]
      summary: 提交审核（发布方）
      operationId: submitArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /artworks/{artworkId}/off-shelf:
    post:
      tags: [Artworks]
      summary: 下架（发布方）
      operationId: offShelfArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/artworks:
    get:
      tags: [Admin]
      summary: 书画审核/平台内容列表
      operationId: adminListArtworksForAudit
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Q"
        - $ref: "#/components/parameters/RegionCode"
        - $ref: "#/components/parameters/AuditStatus"
        - $ref: "#/components/parameters/ArtworkStatus"
        - $ref: "#/components/parameters/ContentSource"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedArtwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: 创建平台自有书画（P1）
      operationId: adminCreateArtwork
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminArtworkCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/artworks/{artworkId}:
    get:
      tags: [Admin]
      summary: 获取书画详情（后台，P1）
      operationId: adminGetArtworkById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin]
      summary: 更新平台自有书画（后台，P1）
      operationId: adminUpdateArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminArtworkUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/artworks/{artworkId}/publish:
    post:
      tags: [Admin]
      summary: 发布书画（后台，P1）
      operationId: adminPublishArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/artworks/{artworkId}/off-shelf:
    post:
      tags: [Admin]
      summary: 下架书画（后台，P1）
      operationId: adminOffShelfArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/artworks/{artworkId}/approve:
    post:
      tags: [Admin]
      summary: 书画审核通过
      operationId: adminApproveArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/artworks/{artworkId}/reject:
    post:
      tags: [Admin]
      summary: 书画审核驳回
      operationId: adminRejectArtwork
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArtworkId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: 关键词（标题/摘要/权利人/发明人/机构名称等）
...
    PriceType:
      name: priceType
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/PriceType"
    PriceMin:
      name: priceMin
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MoneyFen"
    PriceMax:
      name: priceMax
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MoneyFen"
...
    DepositMin:
      name: depositMin
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MoneyFen"
    DepositMax:
      name: depositMax
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MoneyFen"
...
    RegionCode:
      name: regionCode
      in: query
      required: false
      schema:
        type: string
        pattern: "^[0-9]{6}$"
        example: "110000"
      description: 行政区划 adcode（6 位字符串）
    ArtworkCategory:
      name: category
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ArtworkCategory"
      description: 书画类别（书法/绘画）
    CalligraphyScript:
      name: calligraphyScript
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/CalligraphyScript"
    PaintingGenre:
      name: paintingGenre
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/PaintingGenre"
    Creator:
      name: creator
      in: query
      required: false
      schema:
        type: string
      description: 作者/创作者
    CreationYearStart:
      name: creationYearStart
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
      description: 创作年份起
    CreationYearEnd:
      name: creationYearEnd
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
      description: 创作年份止
...
    ArtworkSortBy:
      name: sortBy
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ArtworkSortBy"
...
    ArtworkStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ArtworkStatus"
    ContentSource:
      name: source
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ContentSource"
      description: 内容来源过滤
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    AuditStatus:
      name: auditStatus
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/AuditStatus"
...
    ArtworkId:
      name: artworkId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    MoneyFen:
      type: integer
      format: int64
      minimum: 0
      description: 金额（人民币分）
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    UserRole:
      type: string
      enum: [buyer, seller, cs, admin]
    VerificationType:
      type: string
      enum: [PERSON, COMPANY, ACADEMY, GOVERNMENT, ASSOCIATION, TECH_MANAGER]
    SupplyType:
      type: string
      enum: [UNIVERSITY, UNIVERSITY_985, UNIVERSITY_211, RESEARCH_INSTITUTE, OTHER]
    VerificationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    PriceType:
      type: string
      enum: [FIXED, NEGOTIABLE]
    AuditStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    ContentSource:
      type: string
      enum: [USER, PLATFORM, ADMIN]
...
    ListingStats:
      type: object
      required: [viewCount, favoriteCount, consultCount]
      properties:
        viewCount:
          type: integer
          format: int64
          minimum: 0
        favoriteCount:
          type: integer
          format: int64
          minimum: 0
        consultCount:
          type: integer
          format: int64
          minimum: 0
        commentCount:
          type: integer
          format: int64
          minimum: 0
...
    PageMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    ArtworkSortBy:
      type: string
      enum: [RECOMMENDED, NEWEST, PRICE_ASC, PRICE_DESC]
...
    ArtworkCategory:
      type: string
      enum: [CALLIGRAPHY, PAINTING]
    CalligraphyScript:
      type: string
      enum: [KAISHU, XINGSHU, CAOSHU, LISHU, ZHUANSHU]
    PaintingGenre:
      type: string
      enum: [FIGURE, LANDSCAPE, BIRD_FLOWER, OTHER]
    ArtworkStatus:
      type: string
      enum: [DRAFT, ACTIVE, OFF_SHELF, SOLD]
...
    ContentMedia:
      type: object
      required: [fileId, type, sort]
      properties:
        fileId:
          $ref: "#/components/schemas/Uuid"
        type:
          type: string
          enum: [IMAGE, VIDEO, FILE]
        sort:
          type: integer
          minimum: 0
        url:
          type: string
          format: uri
          readOnly: true
        mimeType:
          type: string
          readOnly: true
        sizeBytes:
          type: integer
          format: int64
          readOnly: true
        fileName:
          type: string
          readOnly: true
...
    ArtworkSummary:
      type: object
      required: [id, title, category, creatorName, priceType, depositAmountFen, auditStatus, status, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        source:
          $ref: "#/components/schemas/ContentSource"
        title:
          type: string
          maxLength: 200
        category:
          $ref: "#/components/schemas/ArtworkCategory"
        calligraphyScript:
          $ref: "#/components/schemas/CalligraphyScript"
        paintingGenre:
          $ref: "#/components/schemas/PaintingGenre"
        creatorName:
          type: string
          maxLength: 200
        creationDate:
          type: string
          format: date
        creationYear:
          type: integer
          minimum: 0
        certificateNo:
          type: string
          maxLength: 100
        priceType:
          $ref: "#/components/schemas/PriceType"
        priceAmountFen:
          $ref: "#/components/schemas/MoneyFen"
        depositAmountFen:
          $ref: "#/components/schemas/MoneyFen"
        regionCode:
          type: string
        material:
          type: string
          maxLength: 200
        size:
          type: string
          maxLength: 100
        coverUrl:
          type: string
          format: uri
        stats:
          $ref: "#/components/schemas/ListingStats"
        auditStatus:
          $ref: "#/components/schemas/AuditStatus"
        status:
          $ref: "#/components/schemas/ArtworkStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Artwork:
      allOf:
        - $ref: "#/components/schemas/ArtworkSummary"
        - type: object
          properties:
            sellerUserId:
              $ref: "#/components/schemas/Uuid"
            description:
              type: string
              maxLength: 5000
            certificateFileIds:
              type: array
              items:
                $ref: "#/components/schemas/Uuid"
            coverFileId:
              $ref: "#/components/schemas/Uuid"
              nullable: true
            media:
              type: array
              items:
                $ref: "#/components/schemas/ContentMedia"
            aiParse:
              $ref: "#/components/schemas/AiParseResult"
              nullable: true
            updatedAt:
              type: string
              format: date-time
    ArtworkPublic:
      allOf:
        - $ref: "#/components/schemas/ArtworkSummary"
        - type: object
          properties:
            seller:
              $ref: "#/components/schemas/UserBrief"
            description:
              type: string
              maxLength: 5000
            media:
              type: array
              items:
                $ref: "#/components/schemas/ContentMedia"
            aiParse:
              $ref: "#/components/schemas/AiParseResult"
              nullable: true
    PagedArtworkSummary:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ArtworkSummary"
        page:
          $ref: "#/components/schemas/PageMeta"
    PagedArtwork:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Artwork"
        page:
          $ref: "#/components/schemas/PageMeta"
    ArtworkCreateRequest:
      type: object
      required: [title, category, creatorName, priceType]
      properties:
        title:
          type: string
          maxLength: 200
        category:
          $ref: "#/components/schemas/ArtworkCategory"
        calligraphyScript:
          $ref: "#/components/schemas/CalligraphyScript"
        paintingGenre:
          $ref: "#/components/schemas/PaintingGenre"
        creatorName:
          type: string
          maxLength: 200
        creationDate:
          type: string
          format: date
        creationYear:
          type: integer
          minimum: 0
        certificateNo:
          type: string
          maxLength: 100
        certificateFileIds:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        description:
          type: string
          maxLength: 5000
        priceType:
          $ref: "#/components/schemas/PriceType"
        priceAmountFen:
          $ref: "#/components/schemas/MoneyFen"
        depositAmountFen:
          $ref: "#/components/schemas/MoneyFen"
        regionCode:
          type: string
        material:
          type: string
          maxLength: 200
        size:
          type: string
          maxLength: 100
        coverFileId:
          $ref: "#/components/schemas/Uuid"
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/ContentMedia"
    ArtworkUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        category:
          $ref: "#/components/schemas/ArtworkCategory"
        calligraphyScript:
          $ref: "#/components/schemas/CalligraphyScript"
        paintingGenre:
          $ref: "#/components/schemas/PaintingGenre"
        creatorName:
          type: string
          maxLength: 200
        creationDate:
          type: string
          format: date
        creationYear:
          type: integer
          minimum: 0
        certificateNo:
          type: string
          maxLength: 100
        certificateFileIds:
          type: array
          items:
            $ref: "#/components/schemas/Uuid"
        description:
          type: string
          maxLength: 5000
        priceType:
          $ref: "#/components/schemas/PriceType"
        priceAmountFen:
          $ref: "#/components/schemas/MoneyFen"
        depositAmountFen:
          $ref: "#/components/schemas/MoneyFen"
        regionCode:
          type: string
        material:
          type: string
          maxLength: 200
        size:
          type: string
          maxLength: 100
        coverFileId:
          $ref: "#/components/schemas/Uuid"
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/ContentMedia"
...
    UserBrief:
      type: object
      required: [id]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
        role:
          $ref: "#/components/schemas/UserRole"
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        verificationType:
          $ref: "#/components/schemas/VerificationType"
          nullable: true
        orgCategory:
          $ref: "#/components/schemas/SupplyType"
...
    AiContentType:
      type: string
      enum: [LISTING, DEMAND, ACHIEVEMENT, ARTWORK]
    AiParseStatus:
      type: string
      enum: [ACTIVE, REVIEW_REQUIRED, REPLACED]
...
    AiParseResult:
      type: object
      required: [id, contentType, contentId, status, confidence, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        contentType:
          $ref: "#/components/schemas/AiContentType"
        contentId:
          $ref: "#/components/schemas/Uuid"
        summaryPlain:
          type: string
        featuresPlain:
          type: string
        keywords:
          type: array
          items:
            type: string
        confidence:
          type: number
          format: double
        modelVersion:
          type: string
        status:
          $ref: "#/components/schemas/AiParseStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminArtworkCreateRequest:
      allOf:
        - $ref: "#/components/schemas/ArtworkCreateRequest"
        - type: object
          properties:
            source:
              $ref: "#/components/schemas/ContentSource"
            sellerUserId:
              $ref: "#/components/schemas/Uuid"
            auditStatus:
              $ref: "#/components/schemas/AuditStatus"
            status:
              $ref: "#/components/schemas/ArtworkStatus"
    AdminArtworkUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/ArtworkUpdateRequest"
        - type: object
          properties:
            source:
              $ref: "#/components/schemas/ContentSource"
            sellerUserId:
              $ref: "#/components/schemas/Uuid"
            auditStatus:
              $ref: "#/components/schemas/AuditStatus"
            status:
              $ref: "#/components/schemas/ArtworkStatus"